Fixed busy waiting due to nonblocking I/O.
Because the default `BufReader` API doesn't allow for a blocking
`read_line` (or any blocking read for that matter), the implementation
would loop in what is essentially a busy wait when noone was writing
to (one of) the FIFOs. This behaviour is now fixed with a workaround
calling out to libc's `poll`.