Implemented timer merging
All timers are now managed by a priority queue, making the number
of threads used for timers constant and not linear in the number
of timers. This change, however, doesn't yet incorporate a more
efficient usage of `select()` to achieve a similar resource
optimization for the FIFO code. Also, some features, like sync,
will need to be reimplemented.